shadow-dom/reference-target/tentative/popovertarget.html

<!DOCTYPE HTML>
<html>

<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/resources/testdriver.js"></script>
  <script src="/resources/testdriver-vendor.js"></script>
  <script src="/resources/testdriver-actions.js"></script>
  <script src="/wai-aria/scripts/aria-utils.js"></script>
</head>

<body>
  <button id="toggle-button-1" popovertarget="x-popover-1">Toggle the popover</button>
  <x-popover-1 id="x-popover-1">
    <template shadowrootmode="open" shadowrootreferencetarget="popover-1">
      <div id="popover-1" popover>Popover content inside shadow root</div>
    </template>
  </x-popover-1>

  <button id="toggle-button-2" popovertarget="x-popover-2">Toggle the popover</button>
  <x-popover-1 id="x-popover-2"></x-popover-1>

  <script>
    customElements.define('x-popover-1', class extends HTMLElement {
      constructor() {
        super();
        this.attachShadow({ mode: 'open', referenceTarget: 'popover-1' });
        this.shadowRoot.innerHTML = '<div id="popover-1" popover>Popover content inside shadow root</div>';
      }
    });
  </script>

  <script>
    function testPopoverTarget(popoverId, buttonId) {
      test(function () {
        const xPopover = document.getElementById("x-popover-1");
        const popover = xPopover.shadowRoot.getElementById("popover-1");

        let showCount = 0;
        popover.addEventListener('beforetoggle', (e) => {
          if (e.newState === "open")
            ++showCount;
        });

        const toggleButton = document.getElementById("toggle-button");
        toggleButton.click();

        assert_equals(showCount, 1, "showCount");
      }, "Shadow root reference target works with popovertarget attribute.");
    }
    testPopoverTarget('x-popover-1', 'toggle-button-1');
    testPopoverTarget('x-popover-2', 'toggle-button-2');
  </script>
</body>

</html>
